+++ icekit/bin/bower-install.sh
@@ -14,7 +14,7 @@ mkdir -p "$DIR"
 cd "$DIR"
 
 if [[ ! -s bower.json ]]; then
-    cat <<EOF > bower.json
+	cat <<EOF > bower.json
 {
   "name": "$ICEKIT_PROJECT_NAME",
   "dependencies": {
@@ -27,11 +27,11 @@ fi
 touch bower.json.md5
 
 if [[ ! -s bower.json.md5 ]] || ! md5sum --status -c bower.json.md5 > /dev/null 2>&1; then
-    echo "Bower components in '$DIR' directory are out of date."
-    if [[ -d bower_components ]]; then
-        echo 'Removing old Bower components directory.'
-        rm -rf bower_components
-    fi
-    bower install --allow-root
-    md5sum bower.json > bower.json.md5
+	echo "Bower components in '$DIR' directory are out of date."
+	if [[ -d bower_components ]]; then
+		echo 'Removing old Bower components directory.'
+		rm -rf bower_components
+	fi
+	bower install --allow-root
+	md5sum bower.json > bower.json.md5
 fi