+++ icekit/project/settings/__init__.py
@@ -1,6 +1,12 @@
 from __future__ import absolute_import
 
 try:
-    from icekit_settings import *
+    try:
+        # Local (override) project settings.
+        from icekit_settings_local import *
+    except ImportError:
+        # Project settings.
+        from icekit_settings import *
 except ImportError:
+    # ICEkit settings.
     from .calculated import *