+++ icekit/tests/models.py
@@ -1,18 +1,23 @@
 """
 Test models for ``icekit`` app.
 """
+from urlparse import urljoin
+
 from django.db import models
 from django.http import HttpResponse
 
 from fluent_pages.extensions import page_type_pool
+from icekit.publishing.models import PublishableFluentContents
 
 from icekit import abstract_models
-from icekit.articles.abstract_models import ListingPage, ArticleBase
-from icekit.articles.page_type_plugins import ListingPagePlugin
+from icekit.content_collections.abstract_models import AbstractListingPage, \
+    AbstractCollectedContent, TitleSlugMixin
+from icekit.content_collections.page_type_plugins import ListingPagePlugin
+
 from icekit.page_types.layout_page.abstract_models import \
     AbstractLayoutPage, AbstractUnpublishableLayoutPage
 from icekit.plugins import ICEkitFluentContentsPagePlugin
-
+from icekit import mixins
 
 class BaseModel(abstract_models.AbstractBaseModel):
     """
@@ -21,15 +26,15 @@ class BaseModel(abstract_models.AbstractBaseModel):
     pass
 
 
-class FooWithLayout(abstract_models.LayoutFieldMixin):
+class FooWithLayout(mixins.LayoutFieldMixin):
     pass
 
 
-class BarWithLayout(abstract_models.LayoutFieldMixin):
+class BarWithLayout(mixins.LayoutFieldMixin):
     pass
 
 
-class BazWithLayout(abstract_models.LayoutFieldMixin):
+class BazWithLayout(mixins.LayoutFieldMixin):
     pass
 
 
@@ -37,14 +42,14 @@ class ImageTest(models.Model):
     image = models.ImageField(upload_to='testing/')
 
 
-class ArticleListing(ListingPage):
+class ArticleListing(AbstractListingPage):
     """A page that lists articles that link to it as parent"""
 
-    def get_items_to_list(self, request):
+    def get_public_items(self, request):
         unpublished_pk = self.get_draft().pk
         return Article.objects.published().filter(parent_id=unpublished_pk)
 
-    def get_items_to_route(self, request):
+    def get_visible_items(self, request):
         unpublished_pk = self.get_draft().pk
         return Article.objects.visible().filter(parent_id=unpublished_pk)
 
@@ -52,7 +57,7 @@ class ArticleListing(ListingPage):
         db_table = 'test_articlelisting'
 
 
-class Article(ArticleBase):
+class Article(AbstractCollectedContent, PublishableFluentContents, TitleSlugMixin):
     """Articles that belong to a particular listing"""
     parent = models.ForeignKey(ArticleListing)
 
@@ -63,6 +68,7 @@ class Article(ArticleBase):
         return HttpResponse(
             u"%s: %s" % (self.parent.get_published().title, self.title)
         )
+
 
 class LayoutPageWithRelatedPages(AbstractLayoutPage):
     related_pages = models.ManyToManyField('fluent_pages.Page')