+++ icekit/tests/tests.py
@@ -7,7 +7,6 @@ import os
 from unittest import skip
 
 from django.apps import apps
-from django.conf import settings
 from django.contrib.auth import get_user_model
 from django.contrib.auth.hashers import make_password
 from django.contrib.contenttypes.models import ContentType
@@ -24,7 +23,7 @@ from fluent_contents.plugins.rawhtml.models import RawHtmlItem
 from fluent_pages.models import PageLayout
 from fluent_pages.pagetypes.fluentpage.models import FluentPage
 from forms_builder.forms.models import Form
-from icekit.abstract_models import LayoutFieldMixin
+from icekit.mixins import LayoutFieldMixin
 from icekit.page_types.layout_page.models import LayoutPage
 from icekit.plugins import descriptors
 from icekit.plugins.faq.models import FAQItem
@@ -40,8 +39,6 @@ from icekit.plugins.twitter_embed.models import TwitterEmbedItem
 from icekit.response_pages.models import ResponsePage
 from mock import patch, Mock
 
-from icekit.tests.models import Article
-from icekit.tests.models import ArticleListing
 from icekit.utils import fluent_contents, implementation
 from icekit.utils.admin import mixins
 from icekit.utils.pagination import describe_page_numbers
@@ -133,7 +130,7 @@ class Forms(WebTest):
         self.assertEqual(teaf.cleaned_data['twitter_url'], twitter_url)
 
 
-class Layout(WebTest):
+class LayoutTest(WebTest):
 
     def test_auto_add(self):
         # No layouts.
@@ -502,6 +499,16 @@ class Views(WebTest):
                 user=self.super_user_1)
             self.assertEqual(response.status_code, 200)
 
+    # Test workaround applied for django-polymorphic 1.0+ per #31
+    def test_can_save_polymorphic_page_in_admin(self):
+        response = self.app.get(
+            reverse('admin:fluent_pages_page_change',
+                    args=(self.layoutpage_1.pk,)),
+            user=self.super_user_1)
+        # Hit 'Save' in form -- does nothing really, but will fail with
+        # `ParentAdminNotRegistered` if polymorphic issue #31 is present.
+        response.forms[0].submit()
+
     def test_response_pages(self):
         response = self.app.get(reverse('404'), expect_errors=404)
         self.assertEqual(response.status_code, 404)
@@ -630,226 +637,3 @@ class TestIceKitTags(WebTest):
         response.mustcontain('<div class="tag-fake-slot"></div>')
         response.mustcontain('<div class="tag-fake-slot-render">None</div>')
         response.mustcontain('div class="filter-fake-slot">None</div>')
-
-
-class TestDescribePageNumbers(TestCase):
-    def test_join_words(self):
-        self.assertEqual(
-            describe_page_numbers(1, 500, 10),
-            {
-                'numbers': [1, 2, 3, 4, None, 48, 49, 50],
-                'has_previous': False,
-                'has_next': True,
-                'previous_page': 0,
-                'current_page': 1,
-                'next_page': 2,
-                'total_count': 500,
-                'per_page': 10,
-            },
-        )
-
-        self.assertEqual(
-            describe_page_numbers(10, 500, 10),
-            {
-                'numbers': [1, 2, 3, None, 7, 8, 9, 10, 11, 12, 13, None, 48, 49, 50],
-                'has_previous': True,
-                'has_next': True,
-                'previous_page': 9,
-                'current_page': 10,
-                'next_page': 11,
-                'total_count': 500,
-                'per_page': 10,
-            },
-        )
-
-        self.assertEqual(
-            describe_page_numbers(50, 500, 10),
-            {
-                'numbers': [1, 2, 3, None, 47, 48, 49, 50],
-                'has_previous': True,
-                'has_next': False,
-                'previous_page': 49,
-                'current_page': 50,
-                'next_page': 51,
-                'total_count': 500,
-                'per_page': 10,
-            },
-        )
-
-
-class TestArticles(WebTest):
-    """
-    Test Article publishing
-    """
-
-    def setUp(self):
-        self.listing_layout = models.Layout.auto_add(
-            'icekit/layouts/listing.html',
-            ArticleListing,
-        )
-
-        # Generate an article listing
-        # Generate an article attached to the listing
-
-        # Don't publish either
-        staff_1 = User.objects.create(
-            email='test@test.com',
-            is_staff=True,
-            is_active=True,
-            is_superuser=True,
-        )
-
-        self.listing = ArticleListing.objects.create(
-            author=staff_1,
-            title="Listing Test",
-            slug="listing-test",
-            layout=self.listing_layout,
-        )
-
-        self.article = Article.objects.create(
-            parent=self.listing,
-            title='Article Test',
-            slug="article-test",
-        )
-
-        self.listing_2 = ArticleListing.objects.create(
-            author=staff_1,
-            title="Listing Test 2",
-            slug="listing-test-2",
-            layout=self.listing_layout,
-        )
-
-        self.article_2 = Article.objects.create(
-            parent=self.listing_2,
-            title='Article Test 2',
-            slug="article-test-2",
-        )
-
-    def test_model(self):
-        self.article.publish()
-        self.article_2.publish()
-        # test the listing contains the published article
-        request = HttpRequest()
-        self.assertTrue(
-            self.article.get_published()
-            in self.listing.get_items_to_list(request))
-        # ...not the draft one
-        self.assertTrue(
-            self.article not in self.listing.get_items_to_list(request))
-        # ...not an article that isn't associated with the listing
-        self.assertTrue(
-            self.article_2 not in self.listing.get_items_to_list(request))
-        self.assertTrue(
-            self.article_2.get_published()
-            not in self.listing.get_items_to_list(request))
-        self.article.unpublish()
-        self.article_2.unpublish()
-
-    def test_urls(self):
-        # Check both have URLs, but 404.
-        self.listing_url = self.listing.get_absolute_url()
-        self.article_url = self.article.get_absolute_url()
-
-        # check articles are at URLs underneath listing
-        self.assertEquals(self.listing_url, '/listing-test/')
-        self.assertEquals(self.article_url, '/listing-test/article-test/')
-
-        # published and draft URLs should be the same
-        self.listing.publish()
-
-        self.article.publish()
-        self.assertEquals(
-            self.listing.get_published().get_absolute_url(),
-            '/listing-test/'
-        )
-        self.assertEquals(
-            self.article.get_published().get_absolute_url(),
-            '/listing-test/article-test/'
-        )
-        self.listing.unpublish()
-        self.article.unpublish()
-
-    def test_article_view(self):
-        # test that article shows if it is published
-        self.listing.publish()
-        self.article.publish()
-
-        self.listing_url = self.listing.get_published().get_absolute_url()
-        self.article_url = self.article.get_published().get_absolute_url()
-
-        response = self.app.get(self.article_url)
-        self.assertEqual(response.status_code, 200)
-        self.assertContains(response, self.listing.title)
-
-        # even if its parent listing is not published
-        # BOO - looks like fluent-pages won't resolve sub-urls of a page
-        # unless the page is published. Leaving commented code to indicate
-        # desired, but not necessary, functionality.
-        # self.article_2.publish()
-        # self.assertFalse(self.listing_2.has_been_published)
-        #
-        # response = self.app.get(self.article_2.get_absolute_url())
-        # self.assertNotContains(response, self.listing_2.title)
-        # self.article_2.unpublish()
-
-        self.listing.unpublish()
-        self.article.unpublish()
-
-    def test_publishing(self):
-        # check URLs 404 by default
-        self.listing_url = self.listing.get_absolute_url()
-        self.article_url = self.article.get_absolute_url()
-        response = self.app.get(self.listing_url, expect_errors=True)
-        self.assertEqual(response.status_code, 404)
-        response = self.app.get(self.article_url, expect_errors=True)
-        self.assertEqual(response.status_code, 404)
-
-        # Publish Article Listing
-        self.listing.publish()
-        self.listing_url = self.listing.get_published().get_absolute_url()
-
-        # Check it works, but doesn't include the Article
-        response = self.app.get(self.listing_url)
-        self.assertEqual(response.status_code, 200)
-
-        self.assertNotContains(response, self.article.title)
-
-        # Publish Article
-        self.article.publish()
-        self.article_url = self.article.get_published().get_absolute_url()
-        response = self.app.get(self.article_url)
-        self.assertEqual(response.status_code, 200)
-
-        # Check the article listing now includes the Article
-        response = self.app.get(
-            self.listing_url)
-        self.assertContains(response, self.article.title)
-
-        # Unpublish Article Listing
-        self.listing.unpublish()
-        response = self.app.get(self.listing_url, expect_errors=True)
-        self.assertEqual(response.status_code, 404)
-
-        # Unpublish Article to tidy up
-        self.article.unpublish()
-        response = self.app.get(self.article_url, expect_errors=True)
-        self.assertEqual(response.status_code, 404)
-
-    def test_listing(self):
-        # Check the listing doesn't include article_2, even if everything is
-        # published
-
-        self.article.publish()
-        self.article_2.publish()
-        self.listing.publish()
-
-        self.listing_url = self.listing.get_published().get_absolute_url()
-        self.article_url = self.article.get_published().get_absolute_url()
-
-        response = self.app.get(self.listing_url)
-        self.assertContains(response, self.article.title)
-        self.assertNotContains(response, self.article_2.title)
-
-        self.article.unpublish()
-        self.article_2.unpublish()
-        self.listing.unpublish()