+++ project_template/articles/migrations/0001_initial.py
@@ -2,8 +2,8 @@
 from __future__ import unicode_literals
 
 from django.db import migrations, models
-import django.db.models.deletion
 import django.utils.timezone
+import django.db.models.deletion
 
 
 class Migration(migrations.Migration):
@@ -17,30 +17,28 @@ class Migration(migrations.Migration):
         migrations.CreateModel(
             name='Article',
             fields=[
-                ('id', models.AutoField(serialize=False, verbose_name='ID', primary_key=True, auto_created=True)),
-                ('publishing_is_draft', models.BooleanField(default=True, editable=False, db_index=True)),
+                ('id', models.AutoField(serialize=False, primary_key=True, auto_created=True, verbose_name='ID')),
+                ('publishing_is_draft', models.BooleanField(db_index=True, editable=False, default=True)),
                 ('publishing_modified_at', models.DateTimeField(default=django.utils.timezone.now, editable=False)),
-                ('publishing_published_at', models.DateTimeField(editable=False, null=True)),
-                ('title', models.CharField(max_length=255)),
-                ('slug', models.SlugField(max_length=255)),
-                ('layout', models.ForeignKey(blank=True, related_name='glamkit_articles_article_related', to='icekit.Layout', null=True)),
+                ('publishing_published_at', models.DateTimeField(null=True, editable=False)),
+                ('layout', models.ForeignKey(to='icekit.Layout', related_name='glamkit_articles_article_related', null=True, blank=True)),
             ],
             options={
                 'abstract': False,
             },
         ),
         migrations.CreateModel(
-            name='ArticleCategory',
+            name='ArticleCategoryPage',
             fields=[
-                ('urlnode_ptr', models.OneToOneField(primary_key=True, parent_link=True, to='fluent_pages.UrlNode', serialize=False, auto_created=True)),
-                ('publishing_is_draft', models.BooleanField(default=True, editable=False, db_index=True)),
+                ('urlnode_ptr', models.OneToOneField(to='fluent_pages.UrlNode', primary_key=True, auto_created=True, parent_link=True, serialize=False)),
+                ('publishing_is_draft', models.BooleanField(db_index=True, editable=False, default=True)),
                 ('publishing_modified_at', models.DateTimeField(default=django.utils.timezone.now, editable=False)),
-                ('publishing_published_at', models.DateTimeField(editable=False, null=True)),
-                ('layout', models.ForeignKey(blank=True, related_name='glamkit_articles_articlecategory_related', to='icekit.Layout', null=True)),
-                ('publishing_linked', models.OneToOneField(related_name='publishing_draft', to='glamkit_articles.ArticleCategory', on_delete=django.db.models.deletion.SET_NULL, null=True, editable=False)),
+                ('publishing_published_at', models.DateTimeField(null=True, editable=False)),
+                ('layout', models.ForeignKey(to='icekit.Layout', related_name='glamkit_articles_articlecategorypage_related', null=True, blank=True)),
+                ('publishing_linked', models.OneToOneField(to='glamkit_articles.ArticleCategoryPage', null=True, editable=False, related_name='publishing_draft', on_delete=django.db.models.deletion.SET_NULL)),
             ],
             options={
-                'db_table': 'pagetype_glamkit_articles_articlecategory',
+                'db_table': 'pagetype_glamkit_articles_articlecategorypage',
                 'abstract': False,
             },
             bases=('fluent_pages.htmlpage', models.Model),
@@ -48,11 +46,11 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='article',
             name='parent',
-            field=models.ForeignKey(blank=True, to='glamkit_articles.ArticleCategory', null=True),
+            field=models.ForeignKey(to='glamkit_articles.ArticleCategoryPage'),
         ),
         migrations.AddField(
             model_name='article',
             name='publishing_linked',
-            field=models.OneToOneField(related_name='publishing_draft', to='glamkit_articles.Article', on_delete=django.db.models.deletion.SET_NULL, null=True, editable=False),
+            field=models.OneToOneField(to='glamkit_articles.Article', null=True, editable=False, related_name='publishing_draft', on_delete=django.db.models.deletion.SET_NULL),
         ),
     ]