+++ project_template/icekit_settings_local.sample.py
@@ -1,0 +1,5 @@
+# This file is ignored by VCS.
+
+from icekit_settings import *
+
+# Override the default project settings.